refactor jurusan index

+++ application/views/jurusan/index.php
@@ -18,7 +18,24 @@
 
     <div class="row">
         <div class="col-6">
-
+            <div class="card">
+                <div class="card-header bg-success text-white">
+                    <h5 class="card-title">Tambah Jurusan</h5>
+                </div>
+                <div class="card-body">
+                    <form action="<?= base_url("jurusan/simpan") ?>" method="POST">
+                        <div class="form-group">
+                            <label for="">Nama jurusan</label>
+                            <input type="text" name="nama_jurusan" placeholder="Masukkan Nama Jurusan" class="form-control">
+                        </div>
+                        <button type="submit" class="btn btn-primary mb-3">
+                            <i class="fas fa-save"></i> Simpan
+                        </button>
+					</form>
+                </div>
+            </div>
+        </div> <!-- End Col -->
+        <div class="col-6">
             <div class="card">
                 <div class="card-header bg-primary text-white">
                     <h5 class="card-title">
@@ -26,7 +43,7 @@
                     </h5>
                 </div>
                 <div class="card-body">
-                    <table class="table table-hover table-stripped table-bordered">
+                    <table class="table table-hover table-stripped table-bordered" id="dataTable">
                         <thead>
                             <tr>
                                 <th>No</th>
@@ -35,7 +52,6 @@
                             </tr>
                         </thead>
                         <tbody>
-
         <?php if (count($jurusan) < 1): ?>
                             <tr>
                                 <td colspan="3" class="text-center h3 p-5">Data Kosong</td>
@@ -51,7 +67,7 @@
                                             <i class="fas fa-edit"></i>
                                         </button>
                                     </a>
-                                    <a href="<?= base_url() ?>jurusan/hapus/<?= $j->id_jurusan ?>">
+                                    <a href="<?= base_url() ?>jurusan/hapus/<?= $j->id_jurusan ?>" onclick="return confirm('Apakah anda ingin menghapus data ini?')">
                                         <button class="btn btn-danger">
                                             <i class="fas fa-trash"></i>
                                         </button>
@@ -63,28 +79,6 @@
                     </table>
                 </div>
             </div>
-
-        </div> <!-- End Col -->
-        <div class="col-6">
-            
-
-            <div class="card">
-                <div class="card-header bg-success text-white">
-                    <h5 class="card-title">Tambah Jurusan</h5>
-                </div>
-                <div class="card-body">
-                    <form action="<?= base_url("jurusan/simpan") ?>" method="POST">
-                        <div class="form-group">
-                            <label for="">Nama jurusan</label>
-                            <input type="text" name="nama_jurusan" class="form-control">
-                        </div>
-                        <button type="submit" class="btn btn-primary mb-3">
-                            <i class="fas fa-save"></i> Simpan
-                        </button>
-					</form>
-                </div>
-            </div>
-        
         </div> <!-- End Col -->
     </div>
